move target string out of risk
