+++ src/vivarium_public_health/risks/__init__.py
@@ -1,5 +1,3 @@
 from .distributions import get_distribution
 from .effect import RiskEffect
 from .base_risk import Risk
-
-from .data_transformations import TargetString
