+++ src/vivarium_public_health/risks/effect.py
@@ -1,7 +1,7 @@
 
-from vivarium_public_health.utilities import EntityString
-from .data_transformations import (TargetString, get_relative_risk_data,
-                                   get_population_attributable_fraction_data, get_exposure_effect)
+from vivarium_public_health.utilities import EntityString, TargetString
+from .data_transformations import (get_relative_risk_data, get_population_attributable_fraction_data,
+                                   get_exposure_effect)
 
 
 class RiskEffect:
