+++ tests/test_utilities.py
@@ -2,7 +2,7 @@ from hypothesis import given
 import hypothesis.strategies as st
 import pytest
 
-from vivarium_public_health.utilities import EntityString
+from vivarium_public_health.utilities import EntityString, TargetString
 
 @st.composite
 def component_string(draw, min_components=0, max_components=None):
@@ -23,3 +23,18 @@ def test_EntityString_pass(s):
     r = EntityString(s)
     assert r.type == entity_type
     assert r.name == entity_name
+
+
+@given(component_string().filter(lambda x: len(x.split('.')) != 3))
+def test_TargetString_fail(s):
+    with pytest.raises(ValueError):
+        TargetString(s)
+
+
+@given(component_string(3, 3))
+def test_TargetString_pass(s):
+    target_type, target_name, target_measure = s.split('.')
+    t = TargetString(s)
+    assert t.type == target_type
+    assert t.name == target_name
+    assert t.measure == target_measure
